src/SVC/requirements.txt
<<<<<<< HEAD
joblib
numpy
joblib
typing
pandas
nltk
regex
scikit-learn
tqdm
python-dotenv
=======
scikit-learn
numpy
joblib
>>>>>>> 9643c8dbf8c807e406368a719a7a61d4b9b6399e
